requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
BareNecessities==0.2.8
bcrypt==3.2.0
bidict==0.23.1
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2020.6.20
chardet==4.0.0
click==8.2.1
cloud-init==25.1.2
cloudpathlib==0.21.1
colorama==0.4.4
command-not-found==0.3
confection==0.1.5
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cymem==2.0.11
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
elastic-transport==8.17.1
elasticsearch==8.18.1
elasticsearch-dsl==8.18.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.3.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Mail==0.10.0
flask-redis==0.4.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
langcodes==3.5.0
language_data==1.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mail==2.1.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==8.10.0
murmurhash==1.0.13
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
openai==1.97.1
packaging==21.3
pexpect==4.8.0
playwright==1.53.0
preshed==3.0.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.1
python-engineio==4.12.2
python-magic==0.4.24
python-socketio==5.13.0
pytz==2022.1
PyYAML==5.4.1
redis==6.2.0
requests==2.25.1
rich==14.0.0
SecretStorage==3.3.1
service-identity==18.1.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smart_open==7.3.0.post1
sniffio==1.3.1
sos==4.8.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
ssh-import-id==5.11
systemd-python==234
thinc==8.3.6
tqdm==4.67.1
Twisted==22.1.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
zipp==1.0.0
zope.interface==5.4.0
flask-jwt-extended
